+++ components/api/Module/RelayHP/RelayHP.h
@@ -11,12 +11,12 @@
 #include "Common.h"
 #include "Module.h"
 
-#if (defined(OI_RELAY_HP)) && defined(MODULE_STANDALONE)
+#if (defined(OI_RELAY_HP))
 #define RelayHP RelayHPStandalone
 #elif (defined(OI_RELAY_HP)) && defined(MODULE_SLAVE)
 #define RelayHP RelayHPSlave
 #elif defined(MODULE_MASTER)
-#define RelayHP RelayHPController
+#define RelayHP RelayHPSlave
 #endif
 
 #if defined(OI_RELAY_HP)
@@ -37,11 +37,11 @@ public:
 
 #if (defined(OI_RELAY_HP)) && defined(MODULE_SLAVE)
 
-#include "SlaveController.h"
+#include "Slave.h"
 #include "RelayCmdHandler.h"
 
 class RelayHPSlave : 
-    public SlaveController, 
+    public Slave, 
     public RelayHPStandalone, 
     public RelayCmdHandler
 {
@@ -55,13 +55,13 @@ public:
 #include "Controller.h"
 #include "RelayCmd.h"
 
-class RelayHPController : 
+class RelayHPSlave : 
     public Controller, 
     public RelayCmd
 {
 public:
 
-    RelayHPController(uint32_t sn = 0) :
+    RelayHPSlave(uint32_t sn = 0) :
         Controller(TYPE_OI_RELAY_HP, sn),
         RelayCmd(this){}