+++ components/middleware/Bus/BusRS.h
@@ -47,7 +47,6 @@ public:
     static void end(void);
     static void write(Frame_t* frame, uint32_t timeout=0);
     static int read(Frame_t* frame, uint32_t timeout=portMAX_DELAY);
-    static int transfer(Frame_t* frame, uint32_t timeout);
 
 private:
 
@@ -55,7 +54,6 @@ private:
     static QueueHandle_t _eventQueue;
     static SemaphoreHandle_t _writeMutex;
     static SemaphoreHandle_t _writeReadMutex;
-    static SemaphoreHandle_t _transferMutex;
 
     static uint8_t _calculateChecksum(Frame_t *frame);
     static bool _verifyChecksum(Frame_t *frame);